+++ product_code/report_service/impl/src/main/scala/de/upb/cs/uc4/report/impl/commands/DeleteReport.scala
@@ -1,0 +1,6 @@
+package de.upb.cs.uc4.report.impl.commands
+
+import akka.actor.typed.ActorRef
+import de.upb.cs.uc4.shared.server.messages.Confirmation
+
+case class DeleteReport(replyTo: ActorRef[Confirmation]) extends ReportCommand