+++ product_code/report_service/impl/src/main/scala/de/upb/cs/uc4/report/impl/commands/SetReport.scala
@@ -1,0 +1,7 @@
+package de.upb.cs.uc4.report.impl.commands
+
+import akka.actor.typed.ActorRef
+import de.upb.cs.uc4.report.impl.actor.Report
+import de.upb.cs.uc4.shared.server.messages.Confirmation
+
+case class SetReport(report: Report, timestamp: String, replyTo: ActorRef[Confirmation]) extends ReportCommand